+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - useLayoutEffect and keep latest state after update (see #20)
 - useForceUpdate uses counter instead of boolean (see #20)
 ### Added
-- New useReduxSelectors
+- New useReduxSelectors (experimental)
 
 ## [1.8.0] - 2019-04-02
 ### Changed